requirements.txt
# Core dependencies
pydantic>=2.0.0
openai>=1.0.0
aiohttp>=3.9.0
pyyaml>=6.0
structlog>=24.0.0
tenacity>=8.2.0

# Data handling
pandas>=2.0.0
openpyxl>=3.1.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
black>=23.0.0
mypy>=1.7.0
ruff>=0.1.0

# Optional for enhanced features
redis>=5.0.0
prometheus-client>=0.19.0
rich>=13.0.0
